instructscience/requirements.txt
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-aliases==1.0.1
defusedxml==0.6.0
diff-match-patch==20200713
divio-cli==3.8.1
Django==2.2
django-ckeditor==6.0.0
django-filer==2.0.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-treebeard==4.3.1
easy-thumbnails==2.7
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pyaml==18.11.0
pytz==2020.1
PyYAML==3.13
requests==2.20.1
six==1.12.0
sqlparse==0.4.1
tablib==2.0.0
tabulate==0.8.3
Unidecode==1.1.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
